Updated Content and Added Form
For this commit, i have added a form for users to signup with, entering their name etc. Additonal changes have been made to the sites UI, how it llooks, adding a footer and changing some colours. I have altered some of the cards to cater for the input of events and more.