Endpoint filtering and update to event details
In this commit I have added event filtering for my db through an end point, similar to my /search endpoint using SQL queries. I copied my additional display layout for my event cards, so when the filter is activated or fetched, it displays the relevant filter. This can be tweaked in later iterations.